pyproject.toml
[tool.poetry]
name = "plane-lazy-bird"
version = "0.1.0"
description = "Plane integration for Lazy-Bird automation"
authors = ["Lazy-Bird Development Team"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/yusufkaraaslan/plane-lazy-bird-integration"
repository = "https://github.com/yusufkaraaslan/plane-lazy-bird-integration"
keywords = ["lazy-bird", "plane", "automation", "django"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
]
packages = [{include = "plane_lazy_bird"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14.0"
httpx = "^0.25.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.5.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.9.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["test_*.py"]
addopts = "--cov=plane_lazy_bird --cov-report=html --cov-report=term-missing"
